+++ jobject-diff/src/test/java/de/alpharogroup/diff/SerializedChangedAttributeResultTest.java
@@ -24,13 +24,16 @@
  */
 package de.alpharogroup.diff;
 
-import static org.testng.Assert.assertTrue;
+import static org.testng.AssertJUnit.assertEquals;
+import static org.testng.AssertJUnit.assertNotSame;
+import static org.testng.AssertJUnit.assertTrue;
 
 import org.testng.annotations.Test;
 
 import de.alpharogroup.diff.beans.SerializedChangedAttributeResult;
 import de.alpharogroup.test.objects.evaluations.EqualsEvaluator;
 import de.alpharogroup.test.objects.evaluations.HashcodeEvaluator;
+import de.alpharogroup.test.objects.evaluations.ToStringEvaluator;
 
 /**
  * The unit test class for the class {@link SerializedChangedAttributeResult}.
@@ -39,15 +42,73 @@ public class SerializedChangedAttributeResultTest
 {
 
 	/**
-	 * Test method for bean {@link SerializedChangedAttributeResult}.
+	 * Test method for {@link SerializedChangedAttributeResult#equals(Object)}
 	 */
 	@Test
-	public void test()
+	public void testEqualsObject()
 	{
-		SerializedChangedAttributeResult changedAttributeResult = SerializedChangedAttributeResult
+		final SerializedChangedAttributeResult expected = SerializedChangedAttributeResult.builder()
+			.attributeName("foo").sourceAttribute("").changedAttribute("").build();
+		final SerializedChangedAttributeResult actual = new SerializedChangedAttributeResult();
+
+		assertNotSame(expected, actual);
+		final SerializedChangedAttributeResult attributeResult = new SerializedChangedAttributeResult();
+		attributeResult.setAttributeName("foo");
+		attributeResult.setSourceAttribute("");
+		attributeResult.setChangedAttribute("");
+		assertEquals(expected, attributeResult);
+		assertTrue(
+			EqualsEvaluator.evaluateReflexivityNonNullSymmetricAndConsistency(expected, actual));
+		assertTrue(EqualsEvaluator.evaluateReflexivityNonNullSymmetricConsistencyAndTransitivity(
+			expected, attributeResult, SerializedChangedAttributeResult.builder()
+				.attributeName("foo").sourceAttribute("").changedAttribute("").build()));
+	}
+
+	/**
+	 * Test method for {@link SerializedChangedAttributeResult#hashCode()}
+	 */
+	@Test
+	public void testHashcode()
+	{
+		boolean expected;
+		boolean actual;
+		final SerializedChangedAttributeResult attributeResult1 = SerializedChangedAttributeResult
 			.builder().attributeName("foo").sourceAttribute("").changedAttribute("").build();
-		assertTrue(HashcodeEvaluator.evaluateConsistency(changedAttributeResult));
-		assertTrue(EqualsEvaluator.evaluateReflexivity(changedAttributeResult));
+		final SerializedChangedAttributeResult attributeResult2 = SerializedChangedAttributeResult
+			.builder().attributeName("foo").sourceAttribute("").changedAttribute("").build();
+		actual = HashcodeEvaluator.evaluateEquality(attributeResult1, attributeResult2);
+		expected = true;
+		assertEquals(expected, actual);
+
+		expected = true;
+		actual = HashcodeEvaluator.evaluateUnequality(attributeResult1,
+			SerializedChangedAttributeResult.builder().attributeName("bar").sourceAttribute("")
+				.changedAttribute("").build());
+		assertEquals(expected, actual);
+
+		actual = HashcodeEvaluator.evaluateConsistency(attributeResult1);
+		expected = true;
+		assertEquals(expected, actual);
+	}
+
+	/**
+	 * Test method for {@link SerializedChangedAttributeResult#toString()}
+	 */
+	@Test
+	public void testToString()
+	{
+		boolean expected;
+		boolean actual;
+		actual = ToStringEvaluator.evaluate(SerializedChangedAttributeResult.class);
+		expected = true;
+		assertEquals(expected, actual);
+
+		final SerializedChangedAttributeResult integerBox = SerializedChangedAttributeResult
+			.builder().attributeName("foo").sourceAttribute("").changedAttribute("").build();
+
+		actual = ToStringEvaluator.evaluateConsistency(integerBox);
+		expected = true;
+		assertEquals(expected, actual);
 	}
 
 }