+++ jobject-diff/src/test/java/de/alpharogroup/diff/object/DiffObjectExtensionsTest.java
@@ -33,6 +33,8 @@ import java.util.Map;
 
 import org.junit.Rule;
 import org.junit.rules.ExpectedException;
+import org.meanbean.test.BeanTestException;
+import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
 import de.alpharogroup.diff.beans.ChangedAttributeResult;
@@ -319,6 +321,17 @@ public class DiffObjectExtensionsTest
 			result, null);
 		assertFalse("Size should be 1 but is " + result.size(), result.size() == 0);
 		assertTrue("Size should be 1 but is " + result.size(), result.size() == 1);
+	}
+
+	/**
+	 * Test method for {@link DiffObjectExtensions} with {@link BeanTester}
+	 */
+	@Test(expectedExceptions = { BeanTestException.class, InvocationTargetException.class,
+			UnsupportedOperationException.class })
+	public void testWithBeanTester()
+	{
+		final BeanTester beanTester = new BeanTester();
+		beanTester.testBean(DiffObjectExtensions.class);
 	}
 
 }